fix: updated ClassroomService to handle exceptions and improved code quality
Changed the return types of createClass, updateClass, and deleteClassroom methods to use Optional and void respectively, to better handle exceptions and improve code quality.
Added a new method assignStudentsToClassroom to assign students to a classroom.
Updated the ClassroomController to handle the new return types and exceptions.
Added error handling for IllegalArgumentException and ClassroomExistsException in the ClassroomController.

+++ src/main/java/com/pshs/ams/app/classrooms/impl/ClassroomServiceImpl.java
@@ -1,13 +1,16 @@
 package com.pshs.ams.app.classrooms.impl;
 
+import com.pshs.ams.app.classrooms.exceptions.ClassroomExistsException;
 import com.pshs.ams.app.classrooms.models.entities.Classroom;
-import com.pshs.ams.global.models.enums.CodeStatus;
 import com.pshs.ams.app.classrooms.services.ClassroomService;
+import com.pshs.ams.app.students.models.entities.Student;
+import com.pshs.ams.global.models.enums.CodeStatus;
 import io.quarkus.panache.common.Page;
 import io.quarkus.panache.common.Sort;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import org.hibernate.Hibernate;
 import org.jboss.logging.Logger;
 
 import java.nio.file.Path;
@@ -43,22 +46,22 @@ public class ClassroomServiceImpl implements ClassroomService {
 	 */
 	@Override
 	@Transactional
-	public CodeStatus createClass(Classroom classroom) {
-		if (classroom.getId() != null) {
-			logger.debug("Class already exists with ID: " + classroom.getId());
-			return CodeStatus.BAD_REQUEST;
+	public Optional<Classroom> createClass(Classroom classroom) throws IllegalArgumentException, ClassroomExistsException {
+		if (classroom == null) {
+			throw new IllegalArgumentException("Classroom cannot be null.");
 		}
 
 		Optional<Classroom> existingClass = Classroom.find("classroomName", classroom.getClassroomName())
-				.firstResultOptional();
+			.firstResultOptional();
 		if (existingClass.isPresent()) {
 			logger.debug("Class already exists: " + existingClass.get().getClassroomName());
-			return CodeStatus.EXISTS;
+			return Optional.empty();
 		}
 
-		classroom.persist();
+		classroom.persistAndFlush();
+//		Classroom.getEntityManager().refresh(classroom); // Get the persisted object
 		logger.debug("Class created: " + classroom.getClassroomName());
-		return CodeStatus.OK;
+		return Optional.of(classroom);
 	}
 
 	/**
@@ -68,18 +71,21 @@ public class ClassroomServiceImpl implements ClassroomService {
 	 */
 	@Override
 	@Transactional
-	public CodeStatus deleteClassroom(Integer id) {
+	public void deleteClassroom(Integer id) throws ClassroomExistsException, IllegalArgumentException {
+		if (id == null || id <= 0) {
+			throw new IllegalArgumentException("Invalid id: " + id);
+		}
+
 		logger.debug("Deleting Class: " + id);
 		Optional<Classroom> existingClass = Classroom.findByIdOptional(id);
 		if (existingClass.isPresent()) {
-			logger.debug("Class found: " + existingClass.get().getClassroomName());
 			existingClass.get().delete();
 			logger.debug("Class deleted: " + existingClass.get().getClassroomName());
-			return CodeStatus.OK;
+			return;
 		}
 
 		logger.debug("Class not found: " + id);
-		return CodeStatus.NOT_FOUND;
+		throw new ClassroomExistsException("Class not found: " + id);
 	}
 
 	/**
@@ -95,10 +101,39 @@ public class ClassroomServiceImpl implements ClassroomService {
 	}
 
 	/**
-	 * @param name
-	 * @param page
-	 * @param sort
-	 * @return
+	 * Assigns a list of students to a specified classroom.
+	 *
+	 * @param classroomId the id of the classroom to which the students will be assigned
+	 * @param students    a list of Student objects to be assigned to the classroom
+	 * @return a list of Student objects that have been successfully assigned to the classroom
+	 * @throws IllegalArgumentException if the classroomId is null or invalid, or if the list of students is null or empty
+	 * @throws ClassroomExistsException if the classroom does not exist
+	 */
+	@Override
+	@Transactional
+	public List<Student> assignStudentsToClassroom(Long classroomId, List<Student> students) throws IllegalArgumentException, ClassroomExistsException {
+		if (classroomId == null || classroomId <= 0 || students == null || students.isEmpty()) {
+			throw new IllegalArgumentException("Invalid classroomId or students");
+		}
+
+		Optional<Classroom> classroomOptional = Classroom.findByIdOptional(classroomId);
+		if (classroomOptional.isEmpty()) {
+			throw new ClassroomExistsException("Classroom does not exist");
+		}
+
+		Classroom classroom = classroomOptional.get();
+		List<Long> studentIds = students.stream().map(Student::getId).toList();
+		Student.update("classroom = ?1 where id in ?2", classroom, studentIds);
+		return students;
+	}
+
+	/**
+	 * Searches for classrooms by name with optional pagination and sorting.
+	 *
+	 * @param name the name to search for
+	 * @param page the pagination object containing pagination parameters (page, size)
+	 * @param sort a sorting object containing sorting parameters (sortBy, sortDirection)
+	 * @return a list of Classroom objects
 	 */
 	@Override
 	public List<Classroom> searchClassroomByName(String name, Page page, Sort sort) {
@@ -119,58 +154,60 @@ public class ClassroomServiceImpl implements ClassroomService {
 
 	@Override
 	@Transactional
-	public CodeStatus updateClass(Classroom classroom) {
+	public Optional<Classroom> updateClass(Classroom classroom) throws IllegalArgumentException {
 		if (classroom.getId() == null) {
-			logger.debug("Classroom ID is null");
-			return CodeStatus.BAD_REQUEST;
+			throw new IllegalArgumentException("Classroom ID is null");
 		}
 
 		Optional<Classroom> existingClass = Classroom.findByIdOptional(classroom.getId());
-		if (existingClass.isPresent()) {
-			Classroom existingClassroom = existingClass.get();
-			boolean updated = false;
-
-			if (classroom.getClassroomName() != null
-					&& !classroom.getClassroomName().equals(existingClassroom.getClassroomName())) {
-				logger.debug("Classroom name updated: " + classroom.getClassroomName());
-				existingClassroom.setClassroomName(classroom.getClassroomName());
-				updated = true;
-			}
-
-			if (classroom.getRoom() != null && !classroom.getRoom().equals(existingClassroom.getRoom())) {
-				logger.debug("Classroom room updated: " + classroom.getRoom());
-				existingClassroom.setRoom(classroom.getRoom());
-				updated = true;
-			}
-
-			if (classroom.getProfilePicture() != null
-					&& !classroom.getProfilePicture().equals(existingClassroom.getProfilePicture())) {
-				logger.debug("Classroom profile picture updated: " + classroom.getProfilePicture());
-				existingClassroom.setProfilePicture(classroom.getProfilePicture());
-				updated = true;
-			}
-
-			if (classroom.getTeacher() != null && !classroom.getTeacher().equals(existingClassroom.getTeacher())) {
-				logger.debug("Classroom teacher updated: " + classroom.getTeacher());
-				existingClassroom.setTeacher(classroom.getTeacher());
-				updated = true;
-			}
-
-			if (classroom.getGradeLevel() != null && !classroom.getGradeLevel().equals(existingClassroom.getGradeLevel())) {
-				logger.debug("Classroom grade level updated: " + classroom.getGradeLevel());
-				existingClassroom.setGradeLevel(classroom.getGradeLevel());
-				updated = true;
-			}
-
-			if (updated) {
-				existingClassroom.setUpdatedAt(Instant.now());
-				existingClassroom.persist();
-				logger.debug("Classroom updated: " + existingClassroom.getId());
-			} else {
-				logger.debug("No changes detected for classroom: " + existingClassroom.getId());
-			}
-			return CodeStatus.OK;
+		if (existingClass.isEmpty()) {
+			logger.debug("Classroom not found: " + classroom.getId());
+			return Optional.empty();
 		}
-		return CodeStatus.NOT_FOUND;
+
+		Classroom existingClassroom = existingClass.get();
+		boolean updated = false;
+		if (classroom.getClassroomName() != null
+			&& !classroom.getClassroomName().equals(existingClassroom.getClassroomName())) {
+			logger.debug("Classroom name updated: " + classroom.getClassroomName());
+			existingClassroom.setClassroomName(classroom.getClassroomName());
+			updated = true;
+		}
+
+		if (classroom.getRoom() != null && !classroom.getRoom().equals(existingClassroom.getRoom())) {
+			logger.debug("Classroom room updated: " + classroom.getRoom());
+			existingClassroom.setRoom(classroom.getRoom());
+			updated = true;
+		}
+
+		if (classroom.getProfilePicture() != null
+			&& !classroom.getProfilePicture().equals(existingClassroom.getProfilePicture())) {
+			logger.debug("Classroom profile picture updated: " + classroom.getProfilePicture());
+			existingClassroom.setProfilePicture(classroom.getProfilePicture());
+			updated = true;
+		}
+
+		if (classroom.getTeacher() != null && !classroom.getTeacher().equals(existingClassroom.getTeacher())) {
+			logger.debug("Classroom teacher updated: " + classroom.getTeacher());
+			existingClassroom.setTeacher(classroom.getTeacher());
+			updated = true;
+		}
+
+		if (classroom.getGradeLevel() != null && !classroom.getGradeLevel().equals(existingClassroom.getGradeLevel())) {
+			logger.debug("Classroom grade level updated: " + classroom.getGradeLevel());
+			existingClassroom.setGradeLevel(classroom.getGradeLevel());
+			updated = true;
+		}
+
+		if (updated) {
+			existingClassroom.setUpdatedAt(Instant.now());
+			existingClassroom.persistAndFlush();
+			logger.debug("Classroom updated: " + existingClassroom.getId());
+			return Optional.of(existingClassroom);
+		} else {
+			logger.debug("No changes detected for classroom: " + existingClassroom.getId());
+		}
+
+		return Optional.empty();
 	}
 }
